+++ src/main/java/org/kainos/ea/resources/SalesEmployeeController.java
@@ -3,6 +3,8 @@ package org.kainos.ea.resources;
 
 import io.swagger.annotations.Api;
 import org.kainos.ea.api.SalesEmployeeService;
+import org.kainos.ea.auth.AccessValidator;
+import org.kainos.ea.cli.Role;
 import org.kainos.ea.cli.SalesEmployeeRequest;
 import org.kainos.ea.client.*;
 import org.kainos.ea.core.SalesEmployeeValidator;
@@ -24,21 +26,29 @@ public class SalesEmployeeController {
     @GET
     @Path("/salesemployee")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllSalesEmployees() {
+    public Response getAllSalesEmployees(@QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(salesEmployeeService.getAllSalesEmployees()).build();
         } catch (FailedToGetSalesEmployeesException e) {
             System.err.println(e.getMessage());
 
             return Response.serverError().build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @GET
     @Path("/salesemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getSalesEmployeeById(@PathParam("id") int id) {
+    public Response getSalesEmployeeById(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(salesEmployeeService.getSalesEmployeeById(id)).build();
         } catch (FailedToGetSalesEmployeeException e) {
             System.err.println(e.getMessage());
@@ -48,14 +58,20 @@ public class SalesEmployeeController {
             System.err.println(e.getMessage());
 
             return Response.status(Response.Status.BAD_REQUEST).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @POST
     @Path("/salesemployee")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createSalesEmployee(SalesEmployeeRequest salesEmployeeRequest) {
+    public Response createSalesEmployee(SalesEmployeeRequest salesEmployeeRequest, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             return Response.ok(salesEmployeeService.createSalesEmployee(salesEmployeeRequest)).build();
         } catch (FailedToCreateSalesEmployeeException e) {
             System.err.println(e.getMessage());
@@ -65,14 +81,20 @@ public class SalesEmployeeController {
             System.err.println(e.getMessage());
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 
     @PUT
     @Path("/salesemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateDeliveryEmployee(@PathParam("id") int id, SalesEmployeeRequest request) {
+    public Response updateDeliveryEmployee(@PathParam("id") int id, SalesEmployeeRequest request, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             salesEmployeeService.updateSalesEmployee(id, request);
 
             return Response.ok().build();
@@ -84,15 +106,20 @@ public class SalesEmployeeController {
             e.printStackTrace();
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
-
     }
 
     @DELETE
     @Path("/salesemployee/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteDeliveryEmployeeId(@PathParam("id") int id) {
+    public Response deleteDeliveryEmployeeId(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            AccessValidator.validateAccess(token, Role.HR);
+
             salesEmployeeService.deleteSalesEmployee(id);
 
             return Response.ok().build();
@@ -104,6 +131,10 @@ public class SalesEmployeeController {
             e.printStackTrace();
 
             return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        } catch (AccessValidationException e) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(e.getMessage()).build();
+        } catch (ForbiddenAccessException e) {
+            return Response.status(Response.Status.FORBIDDEN).entity(e.getMessage()).build();
         }
     }
 }
